Fix remove broken symlink in windows.
