+++ tests/FileHelperTest.php
@@ -233,78 +233,6 @@ final class FileHelperTest extends FileSystemTestCase
         $this->assertEquals('\\\\server/share/path/file', FileHelper::normalizePath('\\\\server\share\path//file'));
     }
 
-    public function testUnlinkSymlink(): void
-    {
-        $dirName = 'unlink';
-        $basePath = $this->testFilePath . '/' . $dirName . '/';
-        $symlinkedFilePath = $basePath . 'symlinks/symlinked-file';
-
-        $this->createFileStructure([
-            $dirName => [
-                'file' => 'Symlinked file.',
-                'symlinks' => [
-                    'symlinked-file' => ['symlink', '../file'],
-                ],
-            ],
-        ]);
-
-        FileHelper::unlink($symlinkedFilePath);
-
-        $this->assertFileDoesNotExist($symlinkedFilePath);
-    }
-
-    /**
-     * 777 gives "read only" flag under Windows
-     *
-     * @see https://github.com/yiisoft/files/issues/21
-     */
-    public function testUnlinkFile777(): void
-    {
-        $dirName = 'unlink';
-        $basePath = $this->testFilePath . '/' . $dirName . '/';
-        $filePath = $basePath . 'file.txt';
-
-        $this->createFileStructure([
-            $dirName => [
-                'file.txt' => 'test',
-            ],
-        ]);
-        chmod($filePath, 777);
-
-        FileHelper::unlink($filePath);
-
-        $this->assertFileDoesNotExist($filePath);
-    }
-
-    public function testUnlinkDirectory(): void
-    {
-        $dirName = 'unlink';
-        $basePath = $this->testFilePath . '/' . $dirName . '/';
-
-        $symlinkedDirectoryPath = $basePath . 'symlinks/symlinked-directory';
-
-        $this->createFileStructure([
-            $dirName => [
-                'directory' => [
-                    'file_in_directory' => 'File in directory.',
-                ],
-                'symlinks' => [
-                    'symlinked-directory' => ['symlink', '../directory'],
-                ],
-            ],
-        ]);
-
-        FileHelper::unlink($symlinkedDirectoryPath);
-
-        $this->assertDirectoryDoesNotExist($symlinkedDirectoryPath);
-    }
-
-    public function testUnlinkNonexistentFile(): void
-    {
-        $this->expectWarning();
-        FileHelper::unlink($this->testFilePath . '/not-exists-file.txt');
-    }
-
     public function testIsEmptyDirectory(): void
     {
         $this->createFileStructure([
